remove old code and refactor node client
