+++ src/core/common.ts
@@ -1,6 +1,6 @@
-import { unique, uniqueBy } from '@/common';
-import sharp from 'sharp';
+import { uniqueBy } from '@/common';
 import { BreachProtocolResult } from './game';
+import { BreachProtocolFragmentResult } from './ocr/base';
 import { Daemon, Sequence } from './sequence';
 
 export const HEX_NUMBERS = ['E9', '1C', 'BD', '55', '7A', 'FF'] as const;
@@ -117,41 +117,10 @@ export class BreachProtocolValidationError extends Error {
 
   constructor(
     public message: string,
-    public readonly data: BreachProtocolRawData
+    public readonly result: BreachProtocolFragmentResult<any, any, any>
   ) {
     super(message);
   }
-}
-
-function validateSymbols(symbols: string[]) {
-  if (!symbols.length) {
-    // [].every(() => {}) returns true
-    return false;
-  }
-
-  return symbols
-    .filter(unique)
-    .every((s) => HEX_NUMBERS.includes(s as HexNumber));
-}
-
-function validateBufferSize(n: number) {
-  return Number.isInteger(n) && n >= BUFFER_SIZE_MIN && n <= BUFFER_SIZE_MAX;
-}
-
-function isSquare(n: number) {
-  return n > 0 && Math.sqrt(n) % 1 === 0;
-}
-
-export function isRawDataValid({
-  grid,
-  daemons,
-  bufferSize,
-}: BreachProtocolRawData) {
-  const isGridValid = validateSymbols(grid) && isSquare(grid.length);
-  const areDaemonsValid = validateSymbols(daemons.flat());
-  const isBufferValid = validateBufferSize(bufferSize);
-
-  return isGridValid && areDaemonsValid && isBufferValid;
 }
 
 export interface BreachProtocolExitStrategy {
@@ -176,38 +145,5 @@ export function resolveExitStrategy(
   return {
     willExit,
     shouldForceClose,
-  };
-}
-
-export const BREACH_PROTOCOL_ASPECT_RATIO = 16 / 9;
-
-/** Return aspect ratio for given resolution and handle edge cases. */
-export function getAspectRatio(x: number, y: number) {
-  // WXGA, very close to 16:9
-  // TODO: test if this resolution correctly ocr buffer size.
-  // https://en.wikipedia.org/wiki/Graphics_display_resolution#WXGA
-  if (y === 768 && (x === 1366 || x === 1360)) {
-    return BREACH_PROTOCOL_ASPECT_RATIO;
-  }
-
-  return x / y;
-}
-
-export function getCroppedBoundingBox(x: number, y: number): sharp.Region {
-  // Resolution with ratio less than one have horizontal black
-  // bars, and ratio greater than one have vertical.
-  // Resolutions with ratio equal to 1 are in 16:9 aspect ratio
-  // and do not require cropping.
-  const ratio = getAspectRatio(x, y) / BREACH_PROTOCOL_ASPECT_RATIO;
-  const width = ratio > 1 ? y * BREACH_PROTOCOL_ASPECT_RATIO : x;
-  const height = ratio < 1 ? x / BREACH_PROTOCOL_ASPECT_RATIO : y;
-  const left = (x - width) / 2;
-  const top = (y - height) / 2;
-
-  return {
-    width,
-    height,
-    left,
-    top,
   };
 }
