+++ src/core/ocr/image-container.ts
@@ -4,7 +4,7 @@ import { BreachProtocolFragmentBoundingBox } from './base';
 export abstract class ImageContainer<T> {
   readonly instance: T;
 
-  readonly dimensions: { width: number; height: number };
+  readonly dimensions: { x: number; y: number };
 
   /** Turn wrapped image into 8 bit black and white fragment. */
   abstract process(
@@ -18,7 +18,8 @@ export abstract class ImageContainer<T> {
 
   abstract toFile(instance: T, fileName: string): Promise<unknown>;
 
-  readonly BREACH_PROTOCOL_ASPECT_RATIO = 16 / 9;
+  /** Aspect ratio of breach protocol. */
+  static readonly ASPECT_RATIO = 16 / 9;
 
   /** Return aspect ratio for given resolution and handle edge cases. */
   getAspectRatio(x: number, y: number) {
@@ -26,21 +27,21 @@ export abstract class ImageContainer<T> {
     // TODO: test if this resolution correctly ocr buffer size.
     // https://en.wikipedia.org/wiki/Graphics_display_resolution#WXGA
     if (y === 768 && (x === 1366 || x === 1360)) {
-      return this.BREACH_PROTOCOL_ASPECT_RATIO;
+      return ImageContainer.ASPECT_RATIO;
     }
 
     return x / y;
   }
 
   getCroppedBoundingBox() {
-    const { width: x, height: y } = this.dimensions;
+    const { x, y } = this.dimensions;
     // Resolution with ratio less than one have horizontal black
     // bars, and ratio greater than one have vertical.
     // Resolutions with ratio equal to 1 are in 16:9 aspect ratio
     // and do not require cropping.
-    const ratio = this.getAspectRatio(x, y) / this.BREACH_PROTOCOL_ASPECT_RATIO;
-    const width = ratio > 1 ? y * this.BREACH_PROTOCOL_ASPECT_RATIO : x;
-    const height = ratio < 1 ? x / this.BREACH_PROTOCOL_ASPECT_RATIO : y;
+    const ratio = this.getAspectRatio(x, y) / ImageContainer.ASPECT_RATIO;
+    const width = ratio > 1 ? y * ImageContainer.ASPECT_RATIO : x;
+    const height = ratio < 1 ? x / ImageContainer.ASPECT_RATIO : y;
     const left = (x - width) / 2;
     const top = (y - height) / 2;
 
@@ -50,10 +51,11 @@ export abstract class ImageContainer<T> {
 
 const SHARP_TOKEN = Symbol('SharpImageContainer');
 
+// NOTE: this class will not work in web environments!
 export class SharpImageContainer extends ImageContainer<sharp.Sharp> {
   constructor(
     public readonly instance: sharp.Sharp,
-    public readonly dimensions: { width: number; height: number },
+    public readonly dimensions: { x: number; y: number },
     token?: Symbol
   ) {
     super();
@@ -68,7 +70,11 @@ export class SharpImageContainer extends ImageContainer<sharp.Sharp> {
   static async create(instance: sharp.Sharp) {
     const { width, height } = await instance.metadata();
 
-    return new SharpImageContainer(instance, { width, height }, SHARP_TOKEN);
+    return new SharpImageContainer(
+      instance,
+      { x: width, y: height },
+      SHARP_TOKEN
+    );
   }
 
   process(threshold: number, boundingBox: BreachProtocolFragmentBoundingBox) {
