+++ src/platform-node/register.ts
@@ -1,16 +1,9 @@
 import { t } from '@/common';
-import { FragmentId } from '@/core';
 import iohook from 'iohook';
 import { options } from './cli';
 import { solveBreachProtocol } from './solve';
 
-export function register({
-  workers,
-  screenId,
-}: {
-  workers: Record<FragmentId, Tesseract.Worker>;
-  screenId: string;
-}) {
+export function register(screenId: string) {
   const ready = t`READY`;
   let isRunning = false;
 
@@ -18,7 +11,7 @@ export function register({
     if (isRunning) return;
     isRunning = true;
 
-    await solveBreachProtocol(workers, screenId);
+    await solveBreachProtocol(screenId);
 
     console.info(ready);
     isRunning = false;
